+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/SearchBuilder.java
@@ -39,6 +39,7 @@ import ca.uhn.fhir.jpa.model.search.StorageProcessingMessage;
 import ca.uhn.fhir.jpa.searchparam.JpaRuntimeSearchParam;
 import ca.uhn.fhir.jpa.searchparam.SearchParameterMap;
 import ca.uhn.fhir.jpa.searchparam.registry.ISearchParamRegistry;
+import ca.uhn.fhir.jpa.searchparam.util.DistanceHelper;
 import ca.uhn.fhir.jpa.util.*;
 import ca.uhn.fhir.model.api.IQueryParameterType;
 import ca.uhn.fhir.model.api.IResource;
@@ -160,7 +161,7 @@ public class SearchBuilder implements ISearchBuilder {
 		theParams.clean();
 
 		// Pull out near-distance first so when it comes time to evaluate near, we already know the distance
-		theParams.setNearDistance(myResourceType);
+		DistanceHelper.setNearDistance(myResourceType, theParams);
 
 		/*
 		 * Check if there is a unique key associated with the set
