+++ hapi-fhir-jpaserver-base/src/test/java/ca/uhn/fhir/jpa/provider/dstu3/ResourceProviderDstu3Test.java
@@ -5,7 +5,6 @@ import ca.uhn.fhir.jpa.dao.data.ISearchDao;
 import ca.uhn.fhir.jpa.entity.Search;
 import ca.uhn.fhir.jpa.provider.r4.ResourceProviderR4Test;
 import ca.uhn.fhir.jpa.search.SearchCoordinatorSvcImpl;
-import ca.uhn.fhir.jpa.util.CoordCalculatorTest;
 import ca.uhn.fhir.model.api.TemporalPrecisionEnum;
 import ca.uhn.fhir.model.primitive.InstantDt;
 import ca.uhn.fhir.model.primitive.UriDt;
@@ -66,7 +65,6 @@ import java.math.BigDecimal;
 import java.net.InetSocketAddress;
 import java.net.Socket;
 import java.net.SocketTimeoutException;
-import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
@@ -4278,141 +4276,6 @@ public class ResourceProviderDstu3Test extends BaseResourceProviderDstu3Test {
 			response.close();
 		}
 
-	}
-
-	// FIXME KHS move distance tests to distance test class
-	@Test
-	public void testNearSearchApproximate() {
-		Location loc = new Location();
-		double latitude = CoordCalculatorTest.LATITUDE_UHN;
-		double longitude = CoordCalculatorTest.LONGITUDE_UHN;
-		Location.LocationPositionComponent position = new Location.LocationPositionComponent().setLatitude(latitude).setLongitude(longitude);
-		loc.setPosition(position);
-		IIdType locId = ourClient.create().resource(loc).execute().getId().toUnqualifiedVersionless();
-
-		{ // In the box
-			double bigEnoughDistance = CoordCalculatorTest.DISTANCE_KM_CHIN_TO_UHN * 2;
-			String url = "/Location?" +
-				Location.SP_NEAR + "=" + CoordCalculatorTest.LATITUDE_CHIN + URLEncoder.encode(":") + CoordCalculatorTest.LONGITUDE_CHIN +
-				"&" +
-				Location.SP_NEAR_DISTANCE + "=" + bigEnoughDistance + URLEncoder.encode("|http://unitsofmeasure.org|km");
-
-			Bundle actual = ourClient
-				.search()
-				.byUrl(ourServerBase + "/" + url)
-				.encodedJson()
-				.prettyPrint()
-				.returnBundle(Bundle.class)
-				.execute();
-
-			assertEquals(1, actual.getEntry().size());
-			assertEquals(locId.getIdPart(), actual.getEntry().get(0).getResource().getIdElement().getIdPart());
-		}
-		{ // Outside the box
-			double tooSmallDistance = CoordCalculatorTest.DISTANCE_KM_CHIN_TO_UHN / 2;
-			String url = "/Location?" +
-				Location.SP_NEAR + "=" + CoordCalculatorTest.LATITUDE_CHIN + URLEncoder.encode(":") + CoordCalculatorTest.LONGITUDE_CHIN +
-				"&" +
-				Location.SP_NEAR_DISTANCE + "=" + tooSmallDistance + URLEncoder.encode("|http://unitsofmeasure.org|km");
-
-			myCaptureQueriesListener.clear();
-			Bundle actual = ourClient
-				.search()
-				.byUrl(ourServerBase + "/" + url)
-				.encodedJson()
-				.prettyPrint()
-				.returnBundle(Bundle.class)
-				.execute();
-			myCaptureQueriesListener.logSelectQueries();
-
-			assertEquals(0, actual.getEntry().size());
-		}
-
-	}
-
-
-	@Test
-	public void testNearSearchDistanceNoDistanceChained() {
-		Location loc = new Location();
-		double latitude = CoordCalculatorTest.LATITUDE_CHIN;
-		double longitude = CoordCalculatorTest.LONGITUDE_CHIN;
-		Location.LocationPositionComponent position = new Location.LocationPositionComponent().setLatitude(latitude).setLongitude(longitude);
-		loc.setPosition(position);
-		IIdType locId = ourClient.create().resource(loc).execute().getId().toUnqualifiedVersionless();
-
-		PractitionerRole pr = new PractitionerRole();
-		pr.addLocation().setReference(locId.getValue());
-		IIdType prId = ourClient.create().resource(pr).execute().getId().toUnqualifiedVersionless();
-
-		String url = "PractitionerRole?location." +
-			Location.SP_NEAR + "=" + latitude + URLEncoder.encode(":") + longitude;
-
-		Bundle actual = ourClient
-			.search()
-			.byUrl(ourServerBase + "/" + url)
-			.encodedJson()
-			.prettyPrint()
-			.returnBundle(Bundle.class)
-			.execute();
-
-		assertEquals(1, actual.getEntry().size());
-		assertEquals(prId.getIdPart(), actual.getEntry().get(0).getResource().getIdElement().getIdPart());
-	}
-
-	@Test
-	public void testNearSearchApproximateChained() {
-		Location loc = new Location();
-		double latitude = CoordCalculatorTest.LATITUDE_UHN;
-		double longitude = CoordCalculatorTest.LONGITUDE_UHN;
-		Location.LocationPositionComponent position = new Location.LocationPositionComponent().setLatitude(latitude).setLongitude(longitude);
-		loc.setPosition(position);
-		myCaptureQueriesListener.clear();
-		IIdType locId = myLocationDao.create(loc).getId().toUnqualifiedVersionless();
-		myCaptureQueriesListener.logInsertQueries();
-
-		PractitionerRole pr = new PractitionerRole();
-		pr.addLocation().setReference(locId.getValue());
-		IIdType prId = myPractitionerRoleDao.create(pr).getId().toUnqualifiedVersionless();
-
-		{ // In the box
-			double bigEnoughDistance = CoordCalculatorTest.DISTANCE_KM_CHIN_TO_UHN * 2;
-			String url = "PractitionerRole?location." +
-				Location.SP_NEAR + "=" + CoordCalculatorTest.LATITUDE_CHIN + URLEncoder.encode(":") + CoordCalculatorTest.LONGITUDE_CHIN +
-				"&" +
-				"location." + Location.SP_NEAR_DISTANCE + "=" + bigEnoughDistance + URLEncoder.encode("|http://unitsofmeasure.org|km");
-
-			myCaptureQueriesListener.clear();
-			Bundle actual = ourClient
-				.search()
-				.byUrl(ourServerBase + "/" + url)
-				.encodedJson()
-				.prettyPrint()
-				.returnBundle(Bundle.class)
-				.execute();
-			myCaptureQueriesListener.logSelectQueries();
-
-			assertEquals(1, actual.getEntry().size());
-			assertEquals(prId.getIdPart(), actual.getEntry().get(0).getResource().getIdElement().getIdPart());
-		}
-		{ // Outside the box
-			double tooSmallDistance = CoordCalculatorTest.DISTANCE_KM_CHIN_TO_UHN / 2;
-			String url = "PractitionerRole?location." +
-				Location.SP_NEAR + "=" + CoordCalculatorTest.LATITUDE_CHIN + URLEncoder.encode(":") + CoordCalculatorTest.LONGITUDE_CHIN +
-				"&" +
-				"location." + Location.SP_NEAR_DISTANCE + "=" + tooSmallDistance + URLEncoder.encode("|http://unitsofmeasure.org|km");
-
-			myCaptureQueriesListener.clear();
-			Bundle actual = ourClient
-				.search()
-				.byUrl(ourServerBase + "/" + url)
-				.encodedJson()
-				.prettyPrint()
-				.returnBundle(Bundle.class)
-				.execute();
-			myCaptureQueriesListener.logSelectQueries();
-
-			assertEquals(0, actual.getEntry().size());
-		}
 	}
 
 	private String toStr(Date theDate) {
