+++ hapi-fhir-jpaserver-base/src/test/java/ca/uhn/fhir/jpa/searchparam/MatchUrlServiceTest.java
@@ -58,7 +58,6 @@ public class MatchUrlServiceTest extends BaseJpaTest {
 				Location.SP_NEAR + "=1000.0:2000.0" +
 				"&" +
 				Location.SP_NEAR_DISTANCE + "=" + kmDistance + "|http://unitsofmeasure.org|km", ourCtx.getResourceDefinition("Location"));
-		map.setNearDistance(Location.class);
 
 		QuantityParam nearDistanceParam = map.getNearDistanceParam();
 		assertEquals(1, map.size());
@@ -75,7 +74,6 @@ public class MatchUrlServiceTest extends BaseJpaTest {
 					"&" +
 					Location.SP_NEAR_DISTANCE + "=2|http://unitsofmeasure.org|km",
 				ourCtx.getResourceDefinition("Location"));
-			map.setNearDistance(Location.class);
 
 			fail();
 		} catch (IllegalArgumentException e) {
@@ -92,7 +90,6 @@ public class MatchUrlServiceTest extends BaseJpaTest {
 					"," +
 					"2|http://unitsofmeasure.org|km",
 				ourCtx.getResourceDefinition("Location"));
-			map.setNearDistance(Location.class);
 
 			fail();
 		} catch (IllegalArgumentException e) {
